fix db error
